v.1.2
Some optimization.

Added Quarantine and Enclose mechanics.

Slightly reworked the discard and removeCard method to account for different piles.

Cleaner summary in City

Fixed some issues with the splash mechanic when drawing from the player deck and when handling outbreaks.

Tried to lower complexity in the Listener.

Added a quicker way to load tests.